# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple ETMS API's for managing expenses by own or by others with groups",
        "title": "ETMS API",
        "termsOfService": "https://swagger.io/terms",
        "contact": {
            "name": "Praveen"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
        "/api/v1/expenses/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new expense based on the user's amount and the preferences",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Expenses"
                ],
                "summary": "Create new expense",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Expenses",
                        "name": "expense",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/expenses.Expenses"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/endpoints.CreatedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/endpoints.BadRequestResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/endpoints.InternalServerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "endpoints.BadRequestResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/endpoints.ErrorMessage"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "endpoints.CreatedResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "endpoints.ErrorMessage": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "endpoints.InternalServerResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/endpoints.ErrorMessage"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "expenses.Expenses": {
            "type": "object",
            "required": [
                "amount",
                "category",
                "created_by",
                "description",
                "title"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "group_expense_split": {
                    "$ref": "#/definitions/expenses.Split"
                },
                "is_group_expense": {
                    "type": "boolean"
                },
                "is_personal_expense": {
                    "type": "boolean"
                },
                "owes_amount": {
                    "type": "string"
                },
                "paid_by": {
                    "type": "string"
                },
                "split_need_to_clear_by": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "expenses.Peer": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "peer_id": {
                    "type": "string"
                }
            }
        },
        "expenses.Split": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "string"
                },
                "involved_peers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/expenses.Peer"
                    }
                },
                "owes_amount": {
                    "type": "string"
                },
                "owes_to": {
                    "type": "string"
                },
                "split_type": {
                    "type": "string"
                }
            }
        },
...
    }
}